CREXDATA System Components/Text Analytics/Web Crawler Component/X-Crawler/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
confluent-kafka==2.8.2
cryptography==41.0.3
debugpy==1.6.7.post1
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fake-headers==1.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kaitaistruct==0.10
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
selenium==4.12.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webdriver-manager==4.0.0
webencodings==0.5.1
wsproto==1.2.0
zipp==3.17.0
zstandard==0.21.0
